backend/requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Migrate==4.0.7
Flask-CORS==4.0.0
Flask-Limiter==3.8.0
Flask-SocketIO==5.3.6
python-socketio==5.11.3
eventlet==0.35.2
redis==5.0.7
celery==5.4.0
python-dotenv==1.0.1
marshmallow==3.21.3
email-validator==2.1.1
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0